src/backend/user/pyproject.toml
[project]
name = "user"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.5",
    "minio>=7.2.12",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.10.2",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.18",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.32.1",
]
[project.scripts]
dev = "uvicorn api.main:app --reload --port 8000"
prod = "uvicorn api.main:app --port 4000 --host 0.0.0.0"